fix(core): validation is http error in base exception filter
BaseExceptionFilter.isHttpError is not validating if statusCode is not a valid http status code.
If for some reason the error contains statusCode that is not valid it will break nestjs to a no recovery point.
